ETradeOAuth/login/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
google-cloud-secret-manager==2.16.4
requests==2.31.0

